+++ src/options/common/value.rs
@@ -1,11 +1,11 @@
-use derive_more::Display;
+use derive_more::{Display, IsVariant};
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::ops::{Deref, DerefMut};
 use std::str::FromStr;
 
 /// Represents a value for some CLI option or config. This exists to support optional values that
 /// have a default value so we can distinguish if a CLI value was a default or explicitly defined.
-#[derive(Copy, Clone, Debug, Display)]
+#[derive(Copy, Clone, Debug, Display, IsVariant)]
 pub enum Value<T> {
     /// Value is a default representation.
     Default(T),