+++ CHANGELOG.md
@@ -7,7 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.20.0-alpha.4]
+## [0.20.0-alpha.5]
+
+### Changed
+
+- CLI subcommands refactored
+  - `distant client select` moved to `distant manager select`
+  - `distant client action` moved to `distant action`
+  - `distant client launch` moved to `distant launch`
+  - `distant client connect` moved to `distant connect`
+  - `distant client lsp` moved to `distant lsp`
+  - `distant client repl` moved to `distant repl`
+  - `distant client shell` moved to `distant shell`
+
+## [0.20.0-alpha.4] - 2023-03-31
 
 ### Added
 
@@ -20,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated a variety of dependencies to latest versions
 
-## [0.20.0-alpha.3]
+## [0.20.0-alpha.3] - 2022-11-27
 
 ### Added
 
@@ -375,7 +388,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   pending upon full channel and no longer locks up
 - stdout, stderr, and stdin of `RemoteProcess` no longer cause deadlock
 
-[Unreleased]: https://github.com/chipsenkbeil/distant/compare/v0.20.0-alpha.4...HEAD
+[Unreleased]: https://github.com/chipsenkbeil/distant/compare/v0.20.0-alpha.5...HEAD
+[0.20.0-alpha.5]: https://github.com/chipsenkbeil/distant/compare/v0.20.0-alpha.4...v0.20.0-alpha.5
 [0.20.0-alpha.4]: https://github.com/chipsenkbeil/distant/compare/v0.20.0-alpha.3...v0.20.0-alpha.4
 [0.20.0-alpha.3]: https://github.com/chipsenkbeil/distant/compare/v0.20.0-alpha.2...v0.20.0-alpha.3
 [0.20.0-alpha.2]: https://github.com/chipsenkbeil/distant/compare/v0.20.0-alpha.1...v0.20.0-alpha.2